sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lattice-sdk"
version = "0.1.0"
description = "Python SDK for Lattice distributed workload scheduler"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Apache-2.0"}
authors = [
    {name = "Lattice Contributors"},
]
keywords = ["lattice", "scheduler", "hpc", "kubernetes", "grpc"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "grpcio>=1.48.0",
    "grpcio-tools>=1.48.0",
    "protobuf>=3.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/witlox/lattice"
Documentation = "https://github.com/witlox/lattice/tree/main/docs"
Repository = "https://github.com/witlox/lattice.git"
Issues = "https://github.com/witlox/lattice/issues"

[tool.ruff]
line-length = 100
target-version = "py38"
select = ["E", "F", "W", "I", "N"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
